requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress Bar
tqdm>=4.62.0

# Optional: Document Generation
python-docx>=0.8.11
python-pptx>=0.6.21
